make CandidateCard a stateless component, show bigger loading bar
