+++ src/pages/RecruiterView/CandidateCard.tsx
@@ -1,28 +1,14 @@
 import Avatar from '@material-ui/core/Avatar';
-import Button from '@material-ui/core/Button';
 import Card from '@material-ui/core/Card';
 import CardContent from '@material-ui/core/CardContent';
 import Chip from '@material-ui/core/Chip';
 import CircularProgress from '@material-ui/core/CircularProgress';
-import Dialog from '@material-ui/core/Dialog';
-import DialogActions from '@material-ui/core/DialogActions';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogTitle from '@material-ui/core/DialogTitle';
-import FormControl from '@material-ui/core/FormControl';
 import Grid from '@material-ui/core/Grid';
-import IconButton from '@material-ui/core/IconButton';
-import Input from '@material-ui/core/Input';
-import InputLabel from '@material-ui/core/InputLabel';
-import Menu from '@material-ui/core/Menu';
-import MenuItem from '@material-ui/core/MenuItem';
-import Select from '@material-ui/core/Select';
 import { makeStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
-import MoreHorizIcon from '@material-ui/icons/MoreHoriz';
-import React, { useState } from 'react';
+import React from 'react';
 
-import { Candidates, CandidateState } from '../../common/types';
-import { saveData } from '../../store';
+import { Candidate } from '../../common/types';
 
 const useStyles = makeStyles(theme => ({
     avatar: {
@@ -54,68 +40,12 @@ const useStyles = makeStyles(theme => ({
     }
 }));
 
-const CandidateCard = ({
-    candidateKey,
-    useCandidatesState
-}: {
-    candidateKey: string;
-    useCandidatesState: [Candidates, React.Dispatch<React.SetStateAction<Candidates>>];
-}) => {
+const CandidateCard = ({ candidateKey, candidate, menu }: { candidateKey: string; candidate: Candidate, menu: JSX.Element | null }) => {
     const classes = useStyles();
-
-    // menu state
-    const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
-
-    // dialog states
-    const [openDialog, setOpenDialog] = useState(false);
-    const [selectedState, setSelectedState] = useState<string>('');
-
-    const [candidatesState, setCandidates] = useCandidatesState;
-    const candidates = { ...candidatesState };
-    const candidate = candidates[candidateKey];
 
     if (!candidate || candidate.deleted) {
         return null;
     }
-
-    if (candidate.state && !selectedState) {
-        setSelectedState(candidate.state);
-    }
-
-    const handleClickMenu = (event: React.MouseEvent<HTMLButtonElement>) => {
-        setAnchorEl(event.currentTarget);
-    };
-
-    const handleCloseMenu = () => {
-        setAnchorEl(null);
-    };
-
-    const handleDelete = (candidateKey: string) => {
-        candidates[candidateKey].deleted = true;
-        saveData(candidates);
-        setCandidates(candidates);
-        handleCloseMenu();
-    };
-
-    const handleChangeState = () => {
-        candidates[candidateKey].state = selectedState;
-        saveData(candidates);
-        setCandidates(candidates);
-        handleCloseChangeStateDialog();
-    };
-
-    const handleChangeSelectedState = (event: React.ChangeEvent<{ value: unknown }>) => {
-        setSelectedState(String(event.target.value) || '');
-    };
-
-    const handleOpenChangeStateDialog = () => {
-        setOpenDialog(true);
-    };
-
-    const handleCloseChangeStateDialog = () => {
-        setOpenDialog(false);
-        handleCloseMenu();
-    };
 
     return (
         <Card key={candidateKey} className={classes.cardRoot}>
@@ -141,55 +71,7 @@ const CandidateCard = ({
                             </Typography>
                             <CircularProgress variant="static" value={candidate.score * 100} />
                         </div>
-                        <IconButton
-                            aria-label="more"
-                            aria-controls="candidate-menu"
-                            aria-haspopup="true"
-                            onClick={handleClickMenu}
-                        >
-                            <MoreHorizIcon />
-                        </IconButton>
-                        <Menu
-                            id="candidate-menu"
-                            anchorEl={anchorEl}
-                            keepMounted
-                            open={Boolean(anchorEl)}
-                            onClose={handleCloseMenu}
-                        >
-                            <MenuItem onClick={handleOpenChangeStateDialog}>Change State</MenuItem>
-                            <Dialog open={openDialog} onClose={handleCloseChangeStateDialog}>
-                                <DialogTitle>Change State</DialogTitle>
-                                <DialogContent>
-                                    <form className={classes.formContainer}>
-                                        <FormControl className={classes.formControl}>
-                                            <InputLabel id="application-state-label">State</InputLabel>
-                                            <Select
-                                                labelId="application-state-label"
-                                                id="application-state-select"
-                                                value={selectedState}
-                                                onChange={handleChangeSelectedState}
-                                                input={<Input />}
-                                            >
-                                                {Object.entries(CandidateState).map((stateKeyValue, key) => {
-                                                    return (
-                                                        <MenuItem key={key} value={stateKeyValue[1]}>
-                                                            {stateKeyValue[1]}
-                                                        </MenuItem>
-                                                    );
-                                                })}
-                                            </Select>
-                                        </FormControl>
-                                    </form>
-                                </DialogContent>
-                                <DialogActions>
-                                    <Button onClick={handleCloseChangeStateDialog}>Cancel</Button>
-                                    <Button onClick={handleChangeState} variant="contained" color="primary">
-                                        Ok
-                                    </Button>
-                                </DialogActions>
-                            </Dialog>
-                            <MenuItem onClick={() => handleDelete(candidateKey)}>Delete</MenuItem>
-                        </Menu>
+                        {menu}
                     </Grid>
                 </Grid>
             </CardContent>
